Renames from measure to comparer class
