+++ depth-comparer.rb
